% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_convert_polygon2_kml.R
\name{convert_polygon2_kml}
\alias{convert_polygon2_kml}
\title{Convert Polygon(s) to KML}
\usage{
convert_polygon2_kml(
  input_dir,
  output_dir = NULL,
  name_field,
  desc_field,
  recursive = FALSE
)
}
\arguments{
\item{input_dir}{Path to the input file. The function is tested for ESRI shapefile, GeoJson, and google's keyhole Markup Language (KML) files}

\item{output_dir}{Path to the directory where the KML files will be saved.}

\item{name_field}{Field name containing names for each polygon.It takes base file name and row number of each polygon.}

\item{desc_field}{Field name containing descriptions for each polygon.}

\item{recursive}{Logical indicating whether to recursively search for files in sub-directories.}
}
\description{
The function transform geospatial data in ESRI shapefile, geojson and keyhold markup languages
files into DJI Pilot2 app accepted kml format.
}
