% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{km_data}
\alias{km_data}
\title{KML Dataset}
\format{
A data frame with 11 observations and 3 variables:
\describe{
  \item{\code{Name}}{Name of the feature.}
  \item{\code{Description}}{Description of the feature.}
  \item{\code{geometry}}{Geometry column.}
}
}
\source{
Test data comes with the packages.
}
\usage{
km_data
}
\description{
This dataset contains spatial data in KML format.
}
\keyword{datasets}
