% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{list_data_types}
\alias{list_data_types}
\title{List types of data available in the package}
\usage{
list_data_types()
}
\value{
A character vector listing the unique file extensions of the test data files.
}
\description{
This function lists the types of data available in the package by examining
the file extensions of the test data files.
}
