% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sh_data}
\alias{sh_data}
\title{Shapefile Dataset}
\format{
A data frame with 11 observations and 5 variables:
\describe{
  \item{\code{Shape_Leng}}{Length of the shape.}
  \item{\code{Shape_Area}}{Area of the shape.}
  \item{\code{tile_id}}{Tile ID for naming.}
  \item{\code{descr}}{Description field.}
  \item{\code{geometry}}{Geometry column.}
}
}
\source{
Test Shapefile data
}
\usage{
sh_data
}
\description{
This dataset contains spatial data in Shapefile format.
}
\keyword{datasets}
