% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape2kml-package.R
\docType{package}
\name{shape2kml-package}
\alias{shape2kml}
\alias{shape2kml-package}
\title{shape2kml: Converts Polygon file or files in a direcoty and convert them into DJI pilot2 app accepted kml format}
\description{
This package has a single function **convert_polygon2_kml**, which reads polygons files eiher "*.shp", "*.json", or "*.kml" format data into DJI Pilot2 compatible kml file.
}
\keyword{internal}
